+++ packages/sol-tracing-utils/src/artifact_adapters/truffle_artifact_adapter.ts
@@ -1,4 +1,4 @@
-import { Compiler, CompilerOptions } from '@0x/sol-compiler';
+import { Compiler, CompilerOptions } from '@dexon-foundation/sol-compiler';
 import * as fs from 'fs';
 import * as glob from 'glob';
 import * as path from 'path';